# api/openapi.yaml
openapi: 3.0.3
info:
  title: Todo List Service API
  description: REST API for managing todo items
  version: 1.0.0

servers:
  - url: http://localhost:8080/api

paths:
  /todos:
    post:
      summary: Create a todo item
      operationId: createTodo
      tags: [ todos ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoItem'
      responses:
        '201':
          description: Todo item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItem'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      summary: List todo items
      operationId: listTodos
      tags: [ todos ]
      parameters:
        - name: status
          in: query
          description: Optional status filter. Currently only 'not done' is supported.
          required: false
          schema:
            type: string
            enum: [ not done ]
      responses:
        '200':
          description: List of todo items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoItem'

  /todos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          example: 1

    get:
      summary: Get a todo item
      operationId: getTodo
      tags: [ todos ]
      responses:
        '200':
          description: Todo item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItem'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update a todo item
      operationId: updateTodo
      tags: [ todos ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  required: false
                status:
                  type: string
                  required: false
                  enum: [ not done, done ]
      responses:
        '200':
          description: Todo item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
components:
  schemas:
    CreateTodoItem:
      type: object
      required: [ description, dueDateTime ]
      properties:
        description:
          type: string
          example: Complete the backend coding challenge
        dueDateTime:
          type: string
          format: date-time
          example: 2026-01-20T15:30:00Z

    TodoItem:
      type: object
      required:
        - id
        - description
        - status
        - creationDateTime
        - dueDateTime
      properties:
        id:
          type: integer
          format: int64
          nullable: true
          example: 1
        description:
          type: string
          example: Complete the backend coding challenge
        status:
          type: string
          enum:
            - not done
            - done
            - past due
          example: not_done
        creationDateTime:
          type: string
          format: date-time
          example: 2026-01-13T08:00:00Z
        dueDateTime:
          type: string
          format: date-time
          example: 2026-01-20T15:30:00Z
        doneDateTime:
          type: string
          format: date-time
          nullable: true
          example: null

    ErrorResponse:
      type: object
      required: [ timestamp, status, error, path ]
      properties:
        timestamp:
          type: string
          format: date-time
          example: 2026-01-13T08:00:00Z
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Todo item with id 123 not found
        path:
          type: string
          example: /api/todos/123

  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    UnprocessableEntity:
      description: Unprocessable Entity - cannot modify past due item
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'